refactoring: Cambie movement por transaction. #222
