+++ src/main/java/trinity/play2learn/backend/economy/transaction/mappers/TransactionMapper.java
@@ -1,8 +1,11 @@
 package trinity.play2learn.backend.economy.transaction.mappers;
 
+import java.util.List;
+
 import trinity.play2learn.backend.activity.activity.models.activity.Activity;
 import trinity.play2learn.backend.admin.subject.models.Subject;
 import trinity.play2learn.backend.economy.reserve.models.Reserve;
+import trinity.play2learn.backend.economy.transaction.dtos.TransactionResponseDto;
 import trinity.play2learn.backend.economy.transaction.models.Transaction;
 import trinity.play2learn.backend.economy.transaction.models.TransactionActor;
 import trinity.play2learn.backend.economy.wallet.models.Wallet;
@@ -30,6 +33,19 @@ public class TransactionMapper {
         .activity(activity)
         .reserve(reserve)
         .build();
+    }
+
+    public static final TransactionResponseDto toDto (Transaction transaction) {
+        return TransactionResponseDto.builder()
+            .amount(Math.round(transaction.getAmount() * 1000.0) / 1000.0)
+            .createdAt(transaction.getCreatedAt())
+            .description(transaction.getDescription())
+            .type((transaction.getOrigin() == TransactionActor.ESTUDIANTE) ? "EGRESO" : "INGRESO")
+            .build();
+    }
+
+    public static final List<TransactionResponseDto> toDtoList (List<Transaction> transactions) {
+        return transactions.stream().map(TransactionMapper::toDto).toList();
     }
  
 }
