+++ src/main/java/trinity/play2learn/backend/economy/wallet/mappers/WalletMapper.java
@@ -2,7 +2,7 @@ package trinity.play2learn.backend.economy.wallet.mappers;
 
 import java.util.List;
 
-import trinity.play2learn.backend.economy.wallet.dtos.response.MovementResponseDto;
+import trinity.play2learn.backend.economy.transaction.dtos.TransactionResponseDto;
 import trinity.play2learn.backend.economy.wallet.dtos.response.WalletCompleteResponseDto;
 import trinity.play2learn.backend.economy.wallet.dtos.response.WalletResponseDto;
 import trinity.play2learn.backend.economy.wallet.models.Wallet;
@@ -17,12 +17,12 @@ public class WalletMapper {
             .build();
     }
 
-    public static WalletCompleteResponseDto toCompleteDto(Wallet wallet, List<MovementResponseDto> movements) {
+    public static WalletCompleteResponseDto toCompleteDto(Wallet wallet, List<TransactionResponseDto> transactions) {
         return WalletCompleteResponseDto.builder()
             .id(wallet.getId())
             .balance(wallet.getBalance())
             .invertedBalance(wallet.getInvertedBalance())
-            .movements(movements)
+            .transactions(transactions)
             .build();
     }
