+++ src/main/java/trinity/play2learn/backend/economy/wallet/services/WalletGetService.java
@@ -8,10 +8,10 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.AllArgsConstructor;
 import trinity.play2learn.backend.admin.student.models.Student;
 import trinity.play2learn.backend.admin.student.services.interfaces.IStudentGetByEmailService;
+import trinity.play2learn.backend.economy.transaction.mappers.TransactionMapper;
 import trinity.play2learn.backend.economy.transaction.models.Transaction;
 import trinity.play2learn.backend.economy.transaction.services.interfaces.ITransactionGetLastTransaccionsService;
 import trinity.play2learn.backend.economy.wallet.dtos.response.WalletCompleteResponseDto;
-import trinity.play2learn.backend.economy.wallet.mappers.MovementMapper;
 import trinity.play2learn.backend.economy.wallet.mappers.WalletMapper;
 import trinity.play2learn.backend.economy.wallet.services.interfaces.IWalletGetService;
 import trinity.play2learn.backend.user.models.User;
@@ -33,7 +33,7 @@ public class WalletGetService implements IWalletGetService {
         
         List<Transaction> transactions = transactionGetLastTransaccionsService.execute(student.getWallet());
     
-        return WalletMapper.toCompleteDto(student.getWallet(), MovementMapper.toDtoList(transactions));
+        return WalletMapper.toCompleteDto(student.getWallet(), TransactionMapper.toDtoList(transactions));
     }
     
 }
