+++ src/main/java/trinity/play2learn/backend/economy/wallet/services/interfaces/IWalletGetLastTransactionsService.java
@@ -1,0 +1,12 @@
+package trinity.play2learn.backend.economy.wallet.services.interfaces;
+
+import java.util.List;
+
+import trinity.play2learn.backend.economy.transaction.dtos.TransactionResponseDto;
+import trinity.play2learn.backend.user.models.User;
+
+public interface IWalletGetLastTransactionsService {
+    
+    public List<TransactionResponseDto> cu65GetLastTransactions (User user);
+    
+}
